pyproject.toml
[project]
name = "bobrai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.8",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
